+++ src/utils/logger.js
@@ -39,13 +39,7 @@ const logger = {
   debug: (msg, data) => log('debug', msg, data),
   info: (msg, data) => log('info', msg, data),
   warn: (msg, data) => log('warn', msg, data),
-  error: (msg, data) => log('error', msg, data),
-  
-  setLevel: (level) => {
-    if (LOG_LEVELS[level] !== undefined) {
-      process.env.LOG_LEVEL = level;
-    }
-  }
+  error: (msg, data) => log('error', msg, data)
 };
 
 module.exports = logger;